+++ visualstudio2017-workload-manageddesktop/tools/ChocolateyInstall.ps1
@@ -1,4 +1,4 @@
-﻿Install-VisualStudioWorkload `
+﻿Add-VisualStudioWorkload `
     -PackageName 'visualstudio2017-workload-manageddesktop' `
     -Workload 'ManagedDesktop' `
     -VisualStudioYear '2017' `