+++ visualstudio2017-workload-manageddesktop/tools/ChocolateyUninstall.ps1
@@ -1,4 +1,4 @@
-﻿Uninstall-VisualStudioWorkload `
+﻿Remove-VisualStudioWorkload `
     -PackageName 'visualstudio2017-workload-manageddesktop' `
     -Workload 'ManagedDesktop' `
     -VisualStudioYear '2017' `