+++ chocolatey-visualstudio.extension/CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## Version 1.0.0
 
-- Initial release with helpers: Install-VisualStudio, Uninstall-VisualStudio, Install-VisualStudioWorkload, Uninstall-VisualStudioWorkload.
+- Initial release with helpers: Install-VisualStudio, Uninstall-VisualStudio, Add-VisualStudioWorkload, Remove-VisualStudioWorkload.
 - Tested with Visual Studio 2017 and 2015, should also work with earlier Visual Studio versions.